requirements.txt
Django==1.11.1
Pillow==4.1.1
lxml==3.3.3
ipython==6.0.0
pyOpenSSL==17.0.0
PyQt4==4.11.4
PyQt5==5.8.2
PySide==1.2.4
Pyrex==0.9.4.1
bpython==0.16
celery==4.0.2
certifi==2017.4.17
cffi==1.10.0
cryptography==1.8.1
cx_Oracle==6.0b2
python_dateutil==2.6.0
django_statsd==2.1.1
django-facebook==6.0.3
django-happenings==0.3.3
docutils==0.13.1
Genshi==0.7
protobuf==3.3.0
java==0.0.1
python_memcached==1.58
mock==2.0.0
ndg_httpsclient==0.4.2
nt==0.0.0
numpy==1.13.0rc2
ordereddict==1.1
org==1.0.0
psycopg2==2.7.1
pyasn1==0.2.3
pyinotify==0.9.6
pylibmc==1.5.2
adhocracy_pysqlite==2.6.3
simplejson==3.10.0
South==1.0.2
sqlparse==0.2.3
tblib==1.3.2
wincertstore==0.2
PyYAML==3.12
